refactoring sidedrawer and main page view models
now one view model called NavigationViewMode
